refactor(voting-election): merge package election into voting
