--- packages/election/README.md
@@ -1,3 +1,0 @@
-# @orbiting/backend-modules-election
-
-Store elections and get them with graphql.
