--- packages/election/graphql/index.js
@@ -1,2 +1,0 @@
-const { loadModule } = require('apollo-modules-node')
-module.exports = loadModule(__dirname)
