--- packages/election/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  ...require('./lib'),
-  graphql: require('./graphql')
-}
