--- packages/election/lib/index.js
@@ -1,1 +1,0 @@
-module.exports = require('export-files')(__dirname)
