--- packages/election/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@orbiting/backend-modules-election",
-  "version": "0.0.0",
-  "description": "store elections and query them from graphql",
-  "main": "index.js",
-  "scripts": {
-    "link": "yarn link"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/orbiting/backends.git"
-  },
-  "author": "Andreas Moor <andreas.moor@republik.ch>",
-  "license": "AGPL-3.0",
-  "publishConfig": {
-    "access": "public"
-  },
-  "bugs": {
-    "url": "https://github.com/orbiting/backends/issues"
-  },
-  "homepage": "https://github.com/orbiting/backends#readme",
-  "devDependencies": {
-    "db-migrate": "^0.11.1",
-    "dotenv": "^5.0.1"
-  },
-  "dependencies": {
-    "apollo-modules-node": "^0.1.4",
-    "export-files": "^2.1.1"
-  }
-}
