+++ packages/voting/README.md
@@ -1,3 +1,3 @@
 # @orbiting/backend-modules-voting
 
-Store voting and get them with graphql.
+Store votings and elections and get them with graphql.
