+++ packages/voting/graphql/resolvers/Candidacy.js
