+++ packages/voting/graphql/resolvers/Election.js
