+++ packages/voting/graphql/resolvers/User.js
