+++ packages/voting/graphql/resolvers/_mutations/cancelCandidacy.js
