+++ packages/voting/graphql/resolvers/_mutations/createElection.js
