+++ packages/voting/graphql/resolvers/_queries/election.js
