+++ packages/voting/graphql/resolvers/_queries/elections.js
