+++ packages/voting/lib/candidacies.js
