+++ packages/voting/lib/db.js
