+++ packages/voting/lib/elections.js
