+++ packages/voting/lib/mail.js
