+++ servers/republik/script/db-migrate-all.js
@@ -15,8 +15,7 @@ const migrationDirs = [
   '../../packages/notifications/migrations',
   '../../packages/access/migrations',
   '../../packages/preview/migrations',
-  '../../packages/voting/migrations',
-  '../../packages/election/migrations'
+  '../../packages/voting/migrations'
 ]
 
 dbMigrateAll(migrationDirs)
